requirements.txt
alembic>=1.13.2,<1.14.0
apispec>=6.6.1,<7.0.0
blinker>=1.8.2,<2.0.0
click>=8.1.7,<9.0.0
Flask>=3.0.3,<4.0.0
Flask-Cors>=4.0.1,<5.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
Flask-Migrate>=4.0.7,<5.0.0
flask-smorest>=0.44.0,<1.0.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
greenlet>=3.0.3,<4.0.0
gunicorn>=23.0.0,<24.0.0
itsdangerous>=2.2.0,<3.0.0
Jinja2>=3.1.4,<4.0.0
Mako>=1.3.5,<2.0.0
MarkupSafe>=2.1.5,<3.0.0
marshmallow>=3.21.3,<4.0.0
packaging>=24.1,<25.0.0
passlib>=1.7.4,<2.0.0
PyJWT>=2.9.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
SQLAlchemy>=2.0.32,<3.0.0
typing_extensions>=4.12.2,<5.0.0
webargs>=8.4.0,<9.0.0
Werkzeug>=3.0.3,<4.0.0
